Add event handler for servings changes
This event handler is activated when the user clicks to increase or
decrease the number of servings.

It uses the function updateServings that receives one argument("dec")
to decrease and any other string to increase.

Also uses the function updateServingsIngredients that receives the
recipe as argument. This changes the count values of the recipe
engredients.